+++ app/src/main/java/com/example/allclear/grade/GradeFragment.java
@@ -1,4 +1,4 @@
-package com.example.allclear;
+package com.example.allclear.grade;
 
 import android.os.Bundle;
 
@@ -9,27 +9,24 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.example.allclear.databinding.FragmentGradeBinding;
-import com.example.allclear.databinding.FragmentMyPageBinding;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link MyPageFragment#newInstance} factory method to
+ * Use the {@link GradeFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class MyPageFragment extends Fragment {
-
+public class GradeFragment extends Fragment {
     private static final String ARG_PARAM1 = "param1";
     private static final String ARG_PARAM2 = "param2";
-
     private String mParam1;
     private String mParam2;
 
-    public MyPageFragment() {
+    public GradeFragment() {
         // Required empty public constructor
     }
 
-    public static MyPageFragment newInstance(String param1, String param2) {
-        MyPageFragment fragment = new MyPageFragment();
+    public static GradeFragment newInstance(String param1, String param2) {
+        GradeFragment fragment = new GradeFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -46,12 +43,12 @@ public class MyPageFragment extends Fragment {
         }
     }
 
-    private FragmentMyPageBinding binding;
+    private FragmentGradeBinding binding;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        binding = FragmentMyPageBinding.inflate(inflater, container, false);
+        binding = FragmentGradeBinding.inflate(inflater, container, false);
         return binding.getRoot();
     }
