+++ app/src/main/java/com/example/allclear/graduation/GraduationFragment.java
@@ -1,4 +1,4 @@
-package com.example.allclear;
+package com.example.allclear.graduation;
 
 import android.os.Bundle;
 
@@ -8,25 +8,27 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.example.allclear.databinding.FragmentGradeBinding;
+import com.example.allclear.databinding.FragmentGraduationBinding;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link GradeFragment#newInstance} factory method to
+ * Use the {@link GraduationFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class GradeFragment extends Fragment {
+public class GraduationFragment extends Fragment {
+
     private static final String ARG_PARAM1 = "param1";
     private static final String ARG_PARAM2 = "param2";
+
     private String mParam1;
     private String mParam2;
 
-    public GradeFragment() {
+    public GraduationFragment() {
         // Required empty public constructor
     }
 
-    public static GradeFragment newInstance(String param1, String param2) {
-        GradeFragment fragment = new GradeFragment();
+    public static GraduationFragment newInstance(String param1, String param2) {
+        GraduationFragment fragment = new GraduationFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -43,12 +45,12 @@ public class GradeFragment extends Fragment {
         }
     }
 
-    private FragmentGradeBinding binding;
+    private FragmentGraduationBinding binding;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        binding = FragmentGradeBinding.inflate(inflater, container, false);
+        binding = FragmentGraduationBinding.inflate(inflater, container, false);
         return binding.getRoot();
     }
