+++ app/src/main/java/com/example/allclear/mypage/MyPageFragment.java
@@ -1,4 +1,4 @@
-package com.example.allclear;
+package com.example.allclear.mypage;
 
 import android.os.Bundle;
 
@@ -8,15 +8,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.example.allclear.databinding.FragmentGradeBinding;
-import com.example.allclear.databinding.FragmentGraduationBinding;
+import com.example.allclear.databinding.FragmentMyPageBinding;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link GraduationFragment#newInstance} factory method to
+ * Use the {@link MyPageFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class GraduationFragment extends Fragment {
+public class MyPageFragment extends Fragment {
 
     private static final String ARG_PARAM1 = "param1";
     private static final String ARG_PARAM2 = "param2";
@@ -24,12 +23,12 @@ public class GraduationFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public GraduationFragment() {
+    public MyPageFragment() {
         // Required empty public constructor
     }
 
-    public static GraduationFragment newInstance(String param1, String param2) {
-        GraduationFragment fragment = new GraduationFragment();
+    public static MyPageFragment newInstance(String param1, String param2) {
+        MyPageFragment fragment = new MyPageFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -46,12 +45,12 @@ public class GraduationFragment extends Fragment {
         }
     }
 
-    private FragmentGraduationBinding binding;
+    private FragmentMyPageBinding binding;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        binding = FragmentGraduationBinding.inflate(inflater, container, false);
+        binding = FragmentMyPageBinding.inflate(inflater, container, false);
         return binding.getRoot();
     }
