+++ app/src/main/java/com/example/allclear/timetable/maketimetable/EssentialSubjectActivity.java
@@ -1,4 +1,4 @@
-package com.example.allclear.maketimetable;
+package com.example.allclear.timetable.maketimetable;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,28 +6,24 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 
-import com.example.allclear.R;
-import com.example.allclear.databinding.ActivitySelectMajorBaseBinding;
-import com.example.allclear.databinding.ActivitySelectSemesterBinding;
+import com.example.allclear.databinding.ActivityEssentialSubjectBinding;
 
-public class SelectMajorBaseActivity extends AppCompatActivity {
-    private ActivitySelectMajorBaseBinding binding;
-
+public class EssentialSubjectActivity extends AppCompatActivity {
+    private ActivityEssentialSubjectBinding binding;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+        binding = ActivityEssentialSubjectBinding.inflate(getLayoutInflater());
         super.onCreate(savedInstanceState);
-        binding = ActivitySelectMajorBaseBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
 
         binding.btnNext.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(SelectMajorBaseActivity.this,SelectEssentialGeneralElectiveActivity.class);
+                Intent intent = new Intent(EssentialSubjectActivity.this, SaveTimeTableActivity.class);
                 startActivity(intent);
             }
         });
-
         binding.btnBack.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
