+++ app/src/main/java/com/example/allclear/timetable/maketimetable/SelectEssentialGeneralElectiveActivity.java
@@ -1,4 +1,4 @@
-package com.example.allclear.maketimetable;
+package com.example.allclear.timetable.maketimetable;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,25 +6,23 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 
-import com.example.allclear.R;
 import com.example.allclear.databinding.ActivitySelectEssentialGeneralElectiveBinding;
-import com.example.allclear.databinding.ActivitySelectMajorBinding;
 
-public class SelectMajorActivity extends AppCompatActivity {
+public class SelectEssentialGeneralElectiveActivity extends AppCompatActivity {
 
-    private ActivitySelectMajorBinding binding;
+    private ActivitySelectEssentialGeneralElectiveBinding binding;
 
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        binding = ActivitySelectMajorBinding.inflate(getLayoutInflater());
+        binding = ActivitySelectEssentialGeneralElectiveBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
 
         binding.btnNext.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(SelectMajorActivity.this,SelectGaneralElectiveActivity.class);
+                Intent intent = new Intent(SelectEssentialGeneralElectiveActivity.this, SelectMajorActivity.class);
                 startActivity(intent);
             }
         });
