+++ app/src/main/java/com/example/allclear/timetable/maketimetable/SelectGaneralElectiveActivity.java
@@ -1,41 +1,33 @@
-package com.example.allclear.maketimetable;
+package com.example.allclear.timetable.maketimetable;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
+import com.example.allclear.databinding.ActivitySelectGaneralElectiveBinding;
 
-import com.example.allclear.R;
-import com.example.allclear.databinding.ActivitySelectSemesterBinding;
-
-public class SelectSemesterActivity extends AppCompatActivity {
-
-    private ActivitySelectSemesterBinding binding;
-
+public class SelectGaneralElectiveActivity extends AppCompatActivity {
+    private ActivitySelectGaneralElectiveBinding binding;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+        binding = ActivitySelectGaneralElectiveBinding.inflate(getLayoutInflater());
         super.onCreate(savedInstanceState);
-        binding = ActivitySelectSemesterBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
 
         binding.btnNext.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(SelectSemesterActivity.this,SelectMajorBaseActivity.class);
+            @Override public void onClick(View v) {
+                Intent intent = new Intent(SelectGaneralElectiveActivity.this, SelfAddOneActivity.class);
                 startActivity(intent);
             }
         });
-
-
         binding.btnBack.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 onBackPressed();
             }
         });
-
     }
 
 }
