+++ app/src/main/java/com/example/allclear/timetable/maketimetable/SelectMajorActivity.java
@@ -1,4 +1,4 @@
-package com.example.allclear.maketimetable;
+package com.example.allclear.timetable.maketimetable;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,25 +6,27 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 
-import com.example.allclear.R;
-import com.example.allclear.databinding.ActivityEssentialSubjectBinding;
+import com.example.allclear.databinding.ActivitySelectMajorBinding;
 
-public class EssentialSubjectActivity extends AppCompatActivity {
-    private ActivityEssentialSubjectBinding binding;
+public class SelectMajorActivity extends AppCompatActivity {
+
+    private ActivitySelectMajorBinding binding;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
-        binding = ActivityEssentialSubjectBinding.inflate(getLayoutInflater());
         super.onCreate(savedInstanceState);
+        binding = ActivitySelectMajorBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
 
         binding.btnNext.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(EssentialSubjectActivity.this, SaveTimeTableActivity.class);
+                Intent intent = new Intent(SelectMajorActivity.this, SelectGaneralElectiveActivity.class);
                 startActivity(intent);
             }
         });
+
         binding.btnBack.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
