+++ app/src/main/java/com/example/allclear/timetable/maketimetable/SelectMajorBaseActivity.java
@@ -1,27 +1,31 @@
-package com.example.allclear.maketimetable;
+package com.example.allclear.timetable.maketimetable;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
-import com.example.allclear.databinding.ActivitySelectGaneralElectiveBinding;
 
-public class SelectGaneralElectiveActivity extends AppCompatActivity {
-    private ActivitySelectGaneralElectiveBinding binding;
+import com.example.allclear.databinding.ActivitySelectMajorBaseBinding;
+
+public class SelectMajorBaseActivity extends AppCompatActivity {
+    private ActivitySelectMajorBaseBinding binding;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
-        binding = ActivitySelectGaneralElectiveBinding.inflate(getLayoutInflater());
         super.onCreate(savedInstanceState);
+        binding = ActivitySelectMajorBaseBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
 
         binding.btnNext.setOnClickListener(new View.OnClickListener() {
-            @Override public void onClick(View v) {
-                Intent intent = new Intent(SelectGaneralElectiveActivity.this, SelfAddOneActivity.class);
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(SelectMajorBaseActivity.this, SelectEssentialGeneralElectiveActivity.class);
                 startActivity(intent);
             }
         });
+
         binding.btnBack.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -29,5 +33,4 @@ public class SelectGaneralElectiveActivity extends AppCompatActivity {
             }
         });
     }
-
 }
