+++ app/src/main/java/com/example/allclear/timetable/maketimetable/SelectSemesterActivity.java
@@ -1,4 +1,4 @@
-package com.example.allclear.maketimetable;
+package com.example.allclear.timetable.maketimetable;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,28 +6,27 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 
-import com.example.allclear.R;
-import com.example.allclear.databinding.ActivitySelectEssentialGeneralElectiveBinding;
-import com.example.allclear.databinding.ActivitySelectMajorBaseBinding;
+import com.example.allclear.databinding.ActivitySelectSemesterBinding;
 
-public class SelectEssentialGeneralElectiveActivity extends AppCompatActivity {
+public class SelectSemesterActivity extends AppCompatActivity {
 
-    private ActivitySelectEssentialGeneralElectiveBinding binding;
+    private ActivitySelectSemesterBinding binding;
 
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        binding = ActivitySelectEssentialGeneralElectiveBinding.inflate(getLayoutInflater());
+        binding = ActivitySelectSemesterBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
 
         binding.btnNext.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(SelectEssentialGeneralElectiveActivity.this,SelectMajorActivity.class);
+                Intent intent = new Intent(SelectSemesterActivity.this, SelectMajorBaseActivity.class);
                 startActivity(intent);
             }
         });
+
 
         binding.btnBack.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -35,5 +34,7 @@ public class SelectEssentialGeneralElectiveActivity extends AppCompatActivity {
                 onBackPressed();
             }
         });
+
     }
+
 }
