+++ app/src/main/java/com/example/allclear/timetable/maketimetable/SelfAddTwoActivity.java
@@ -1,4 +1,4 @@
-package com.example.allclear.maketimetable;
+package com.example.allclear.timetable.maketimetable;
 
 import androidx.appcompat.app.AppCompatActivity;
