+++ app/src/main/res/layout/activity_essential_subject.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".maketimetable.EssentialSubjectActivity">
+    tools:context=".timetable.maketimetable.EssentialSubjectActivity">
 
     <TextView
         android:id="@+id/tv_title"
