+++ app/src/main/res/layout/activity_select_ganeral_elective.xml
@@ -4,12 +4,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".maketimetable.SelectGaneralElectiveActivity">
+    tools:context=".timetable.maketimetable.SelectGaneralElectiveActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".maketimetable.SelectEssentialGeneralElectiveActivity">
+        tools:context=".timetable.maketimetable.SelectEssentialGeneralElectiveActivity">
 
         <TextView
             android:id="@+id/tv_title"
