+++ app/src/main/res/layout/activity_select_semester.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".maketimetable.SelectSemesterActivity">
+    tools:context=".timetable.maketimetable.SelectSemesterActivity">
 
     <TextView
         android:id="@+id/tv_title"
