+++ app/src/main/res/layout/activity_self_add_one.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".maketimetable.SelfAddOneActivity">
+    tools:context=".timetable.maketimetable.SelfAddOneActivity">
 
     <TextView
         android:id="@+id/tv_title"
